refactor: rename to curve-common
