+++ apps/lend/src/layout/Header.tsx
@@ -11,7 +11,7 @@ import { getNetworkFromUrl, getRestFullPathname } from '@/utils/utilsRouter'
 import { getParamsFromUrl, getRestPartialPathname } from '@/utils/utilsRouter'
 import { getWalletSignerAddress } from '@/store/createWalletSlice'
 import { _parseRouteAndIsActive, FORMAT_OPTIONS, formatNumber, isLoading } from '@/ui/utils'
-import { useConnectWallet } from '@/shared/features/connect-wallet'
+import { useConnectWallet } from '@/common/shared/features/connect-wallet'
 import { useHeightResizeObserver } from '@/ui/hooks'
 import { visibleNetworksList } from '@/networks'
 import networks from '@/networks'
