+++ apps/lend/src/layout/index.tsx
@@ -6,7 +6,7 @@ import { layoutHeightKeys } from '@/store/createLayoutSlice'
 import { getNetworkFromUrl } from '@/utils/utilsRouter'
 import { getWalletChainId } from '@/store/createWalletSlice'
 import { isFailure, isLoading } from '@/ui/utils'
-import { useConnectWallet } from '@/shared/features/connect-wallet'
+import { useConnectWallet } from '@/common/shared/features/connect-wallet'
 import useStore from '@/store/useStore'
 
 import Header from '@/layout/Header'
