+++ apps/loan/package.json
@@ -30,7 +30,7 @@
     "@supercharge/promise-pool": "^2.3.2",
     "bignumber.js": "^9.0.1",
     "cross-fetch": "^3.1.5",
-    "curve-features": "*",
+    "curve-common": "*",
     "curve-lib": "*",
     "dayjs": "^1.11.7",
     "ethers": "^5",
