+++ apps/main/src/hooks/usePageOnMount.ts
@@ -4,7 +4,7 @@ import type { INetworkName } from '@curvefi/api/lib/interfaces'
 
 import { ethers } from 'ethers'
 import { useCallback, useEffect } from 'react'
-import { useConnectWallet, useSetChain, useSetLocale } from '@/shared/features/connect-wallet'
+import { useConnectWallet, useSetChain, useSetLocale } from '@/common/shared/features/connect-wallet'
 
 import { CONNECT_STAGE, REFRESH_INTERVAL, ROUTE } from '@/constants'
 import { dynamicActivate, updateAppLocale } from '@/lib/i18n'
