+++ packages/curve-common/global.d.ts
