+++ packages/curve-common/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "onboard-helpers",
+  "name": "curve-common",
   "version": "1.0.2",
   "main": "src/index.ts",
   "types": "src/index.ts",
