+++ packages/curve-common/src/connect-wallet/lib/init.ts
