+++ packages/curve-common/src/connect-wallet/lib/locales/index.ts
