+++ packages/curve-common/src/connect-wallet/lib/locales/zh-Hans.ts
