+++ packages/curve-common/src/connect-wallet/lib/locales/zh-Hant.ts
