+++ packages/curve-common/src/connect-wallet/lib/utils/walletModules.ts
