+++ packages/curve-common/tsconfig.json
